KCL literals have their own type system instead of reusing JSON.
We now distinguish between KCL literals of:
- String
- f64
- i64
- u64

instead of reusing JSON values as the only KCL literal.

+++ src/wasm-lib/kcl/src/parser/math.rs
@@ -94,7 +94,7 @@ mod tests {
     #[test]
     fn parse_and_evaluate() {
         /// Make a literal
-        fn lit(n: u8) -> BinaryPart {
+        fn lit(n: u64) -> BinaryPart {
             BinaryPart::Literal(Box::new(Literal {
                 start: 0,
                 end: 0,
